+++ app/models/unit.rb
@@ -36,7 +36,8 @@ class Unit < ApplicationRecord
       :download_jplag_report,
       :get_marking_sessions,
       :get_tutor_times,
-      :grant_extensions
+      :grant_extensions,
+      :get_staff_notes
     ]
 
     # What can convenors do with units?
@@ -66,7 +67,8 @@ class Unit < ApplicationRecord
       :get_tutor_times_summary,
       :get_marking_sessions,
       :upload_grades_csv,
-      :grant_extensions
+      :grant_extensions,
+      :get_staff_notes
     ]
 
     # What can admin do with units?
@@ -90,7 +92,8 @@ class Unit < ApplicationRecord
       :get_feedback_chips,
       :grant_spec_con,
       :download_jplag_report,
-      :get_marking_sessions
+      :get_marking_sessions,
+      :get_staff_notes
     ]
 
     # What can auditors do with units?
@@ -1830,6 +1833,36 @@ class Unit < ApplicationRecord
             result << row["people_#{td.id}"] if td.is_group_task?
             result
           end.flatten)
+        end
+    end
+  end
+
+  def staff_notes_csv
+    CSV.generate() do |csv|
+      # Add headers
+      csv << ([
+        'Student Username',
+        'Project ID',
+        'Student Name',
+        'Staff Note',
+        'Created',
+        'Author Name',
+        'Author Username',
+      ])
+
+      StaffNote.joins(project: :unit)
+        .where(units: { id: id })
+        .order('projects.id', 'created_at')
+        .each do |row|
+          csv << [
+            row.project.student.username,
+            row.project.id.to_s,
+            row.project.student.name,
+            row.note,
+            row.created_at.localtime,
+            row.user.name,
+            row.user.username,
+          ]
         end
     end
   end